cas_project/requirements.txt
numpy>=1.24.0
networkx>=3.1
matplotlib>=3.7.0
pyyaml>=6.0
jinja2>=3.1
fpdf2>=2.7
pytest>=7.3.0
